% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochgen.R
\name{brownian_matrix}
\alias{brownian_matrix}
\title{Simulate a bundle of Brownian Motion paths}
\usage{
brownian_matrix(dt, n_steps, n_traj, x0 = 0, mu = 0, sigma = 1)
}
\arguments{
\item{dt}{Time step size}

\item{n_steps}{Number of steps in each path}

\item{n_traj}{Number of simulated paths}

\item{x0}{Initial value}

\item{mu}{Drift coefficient}

\item{sigma}{Volatility coefficient}
}
\value{
Numeric matrix of dimension \code{n_steps x n_traj} containing simulated paths.
Each column corresponds to one trajectory.
}
\description{
Simulates \code{n_traj} independent brownian motion paths following the SDE:
\deqn{dS_t = \mu dt + \sigma dW_t}
With solution
\deqn{S_t = \mu t + \sigma W_t}
where \eqn{W_t} is the Wiener process.
}
\examples{
# Simulate a bundle of Brownian Motion paths
M <- brownian_matrix(dt = 0.01, n_steps = 100, n_traj = 10)
matplot(M, type = "l", lty = 1,
        xlab = "step", ylab="S_t",
        main = "Brownian Motion")
}
\references{
\href{https://en.wikipedia.org/wiki/Wiener_process}{Wiener process}
}
\seealso{
\code{\link[=brownian_vector]{brownian_vector()}}, \code{\link[=brownian_end_info]{brownian_end_info()}},
}
