% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochgen.R
\name{geom_brownian_matrix}
\alias{geom_brownian_matrix}
\title{TODO}
\usage{
geom_brownian_matrix(dt, n_steps, n_traj, x0 = 1, mu = 0, sigma = 1)
}
\arguments{
\item{TODO}{}
}
\description{
TODO
}
\seealso{
TODO
}
