% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochgen.R
\name{ornstein_uhlenbeck_end_info}
\alias{ornstein_uhlenbeck_end_info}
\title{TODO}
\usage{
ornstein_uhlenbeck_end_info(
  dt,
  n_steps,
  n_traj,
  x0 = 0,
  mu = 0,
  theta = 1,
  sigma = 1,
  lower_thresh = -1,
  upper_thresh = 1
)
}
\arguments{
\item{TODO}{}
}
\description{
TODO
}
\seealso{
TODO
}
