% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochgen.R
\name{ornstein_uhlenbeck_vector}
\alias{ornstein_uhlenbeck_vector}
\title{TODO}
\usage{
ornstein_uhlenbeck_vector(dt, n, x0 = 0, mu = 0, theta = 1, sigma = 1)
}
\arguments{
\item{TODO}{}
}
\description{
TODO
}
\seealso{
TODO
}
