% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochgen.R
\name{poiss_jump_times}
\alias{poiss_jump_times}
\title{Generete a vector of jump times of a Poisson process}
\usage{
poiss_jump_times(lambda, n)
}
\description{
A Poisson process with intensity \eqn{\lambda} can be constructed by counting jumps
which occur at random such that the time between jumps
follows the Exponential distribution with parameter \lambda
}
