% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochgen.R
\docType{package}
\name{stochgen-package}
\alias{stochgen}
\alias{stochgen-package}
\title{stochgen: Simulation of Stochastic Processes}
\description{
A lightweight package for simulating stochastic processes.
}
\author{
\strong{Maintainer}: Antoni Kobosz \email{antoni.kobosz@gmail.com}

}
