% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochgen.R
\name{brownian_end_info}
\alias{brownian_end_info}
\title{TODO}
\usage{
brownian_end_info(x0, mu, sigma, dt, lower_thresh, upper_thresh, n, m)
}
\arguments{
\item{TODO}{}
}
\description{
TODO
}
\seealso{
TODO
}
