% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochgen.R
\name{geom_brownian_end_info}
\alias{geom_brownian_end_info}
\title{TODO}
\usage{
geom_brownian_end_info(
  dt,
  n_steps,
  n_traj,
  x0 = 1,
  mu = 0,
  sigma = 1,
  lower_thresh,
  upper_thresh
)
}
\arguments{
\item{TODO}{}
}
\description{
TODO
}
\seealso{
TODO
}
